Network.AWS.ServiceCatalog.AcceptPortfolioShare

Description

Accepts an offer to share a portfolio.

Creating a Request

acceptPortfolioShare

Arguments

:: Text
apsPortfolioId

-> AcceptPortfolioShare

Creates a value of AcceptPortfolioShare with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apsAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

apsPortfolioId - The portfolio identifier.

data AcceptPortfolioShare

See: acceptPortfolioShare smart constructor.

Request Lenses

apsAcceptLanguage :: Lens' AcceptPortfolioShare (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

apsPortfolioId :: Lens' AcceptPortfolioShare Text

The portfolio identifier.

Destructuring the Response

acceptPortfolioShareResponse

Arguments

:: Int
apsrsResponseStatus

-> AcceptPortfolioShareResponse

Creates a value of AcceptPortfolioShareResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apsrsResponseStatus - -- | The response status code.

data AcceptPortfolioShareResponse

See: acceptPortfolioShareResponse smart constructor.

Response Lenses

apsrsResponseStatus :: Lens' AcceptPortfolioShareResponse Int

- | The response status code.