Network.AWS.ServiceCatalog.CreateConstraint

Description

Creates a new constraint. For more information, see Using Constraints .

Creating a Request

createConstraint

Arguments

:: Text
ccPortfolioId

-> Text
ccProductId

-> Text
ccParameters

-> Text
ccType

-> Text
ccIdempotencyToken

-> CreateConstraint

Creates a value of CreateConstraint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

ccDescription - The text description of the constraint.

ccPortfolioId - The portfolio identifier.

ccProductId - The product identifier.

ccParameters - The constraint parameters. Expected values vary depending on which Type is specified. For examples, see the bottom of this topic. For Type LAUNCH , the RoleArn property is required. For Type NOTIFICATION , the NotificationArns property is required. For Type TEMPLATE , the Rules property is required.

ccType - The type of the constraint. Case-sensitive valid values are: LAUNCH , NOTIFICATION , or TEMPLATE .

ccIdempotencyToken - A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

data CreateConstraint

See: createConstraint smart constructor.

Request Lenses

ccAcceptLanguage :: Lens' CreateConstraint (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

ccDescription :: Lens' CreateConstraint (Maybe Text)

The text description of the constraint.

ccPortfolioId :: Lens' CreateConstraint Text

The portfolio identifier.

ccProductId :: Lens' CreateConstraint Text

The product identifier.

ccParameters :: Lens' CreateConstraint Text

The constraint parameters. Expected values vary depending on which Type is specified. For examples, see the bottom of this topic. For Type LAUNCH , the RoleArn property is required. For Type NOTIFICATION , the NotificationArns property is required. For Type TEMPLATE , the Rules property is required.

ccType :: Lens' CreateConstraint Text

The type of the constraint. Case-sensitive valid values are: LAUNCH , NOTIFICATION , or TEMPLATE .

ccIdempotencyToken :: Lens' CreateConstraint Text

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

Destructuring the Response

createConstraintResponse

Arguments

:: Int
ccrsResponseStatus

-> CreateConstraintResponse

Creates a value of CreateConstraintResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsStatus - The status of the current request.

ccrsConstraintDetail - The resulting detailed constraint information.

ccrsConstraintParameters - The resulting constraint parameters.

ccrsResponseStatus - -- | The response status code.

data CreateConstraintResponse

See: createConstraintResponse smart constructor.

Response Lenses

ccrsStatus :: Lens' CreateConstraintResponse (Maybe RequestStatus)

The status of the current request.

ccrsConstraintDetail :: Lens' CreateConstraintResponse (Maybe ConstraintDetail)

The resulting detailed constraint information.

ccrsConstraintParameters :: Lens' CreateConstraintResponse (Maybe Text)

The resulting constraint parameters.

ccrsResponseStatus :: Lens' CreateConstraintResponse Int

- | The response status code.