Network.AWS.ServiceCatalog.CreateTagOption

Description

Create a new TagOption.

Creating a Request

createTagOption

Arguments

:: Text
ctoKey

-> Text
ctoValue

-> CreateTagOption

Creates a value of CreateTagOption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctoKey - The TagOption key.

ctoValue - The TagOption value.

data CreateTagOption

See: createTagOption smart constructor.

Request Lenses

ctoKey :: Lens' CreateTagOption Text

The TagOption key.

ctoValue :: Lens' CreateTagOption Text

The TagOption value.

Destructuring the Response

createTagOptionResponse

Arguments

:: Int
ctorsResponseStatus

-> CreateTagOptionResponse

Creates a value of CreateTagOptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctorsTagOptionDetail - The resulting detailed TagOption information.

ctorsResponseStatus - -- | The response status code.

data CreateTagOptionResponse

See: createTagOptionResponse smart constructor.

Response Lenses

ctorsTagOptionDetail :: Lens' CreateTagOptionResponse (Maybe TagOptionDetail)

The resulting detailed TagOption information.

ctorsResponseStatus :: Lens' CreateTagOptionResponse Int

- | The response status code.