Network.AWS.ServiceCatalog.DeleteConstraint

Description

Deletes the specified constraint.

Creating a Request

deleteConstraint

Arguments

:: Text
dcId

-> DeleteConstraint

Creates a value of DeleteConstraint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dcId - The identifier of the constraint to delete.

data DeleteConstraint

See: deleteConstraint smart constructor.

Request Lenses

dcAcceptLanguage :: Lens' DeleteConstraint (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dcId :: Lens' DeleteConstraint Text

The identifier of the constraint to delete.

Destructuring the Response

deleteConstraintResponse

Arguments

:: Int
dcrsResponseStatus

-> DeleteConstraintResponse

Creates a value of DeleteConstraintResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsResponseStatus - -- | The response status code.

data DeleteConstraintResponse

See: deleteConstraintResponse smart constructor.

Response Lenses

dcrsResponseStatus :: Lens' DeleteConstraintResponse Int

- | The response status code.