Network.AWS.ServiceCatalog.DescribeConstraint

Description

Retrieves detailed information for a specified constraint.

Creating a Request

describeConstraint

Arguments

:: Text
dId

-> DescribeConstraint

Creates a value of DescribeConstraint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dId - The identifier of the constraint.

data DescribeConstraint

See: describeConstraint smart constructor.

Request Lenses

dAcceptLanguage :: Lens' DescribeConstraint (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dId :: Lens' DescribeConstraint Text

The identifier of the constraint.

Destructuring the Response

describeConstraintResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeConstraintResponse

Creates a value of DescribeConstraintResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsStatus - The status of the current request.

desrsConstraintDetail - Detailed constraint information.

desrsConstraintParameters - The current parameters associated with the specified constraint.

desrsResponseStatus - -- | The response status code.

data DescribeConstraintResponse

See: describeConstraintResponse smart constructor.

Response Lenses

desrsStatus :: Lens' DescribeConstraintResponse (Maybe RequestStatus)

The status of the current request.

desrsConstraintDetail :: Lens' DescribeConstraintResponse (Maybe ConstraintDetail)

Detailed constraint information.

desrsConstraintParameters :: Lens' DescribeConstraintResponse (Maybe Text)

The current parameters associated with the specified constraint.

desrsResponseStatus :: Lens' DescribeConstraintResponse Int

- | The response status code.