Network.AWS.ServiceCatalog.DescribePortfolio

Description

Retrieves detailed information and any tags associated with the specified portfolio.

Creating a Request

describePortfolio

Arguments

:: Text
desId

-> DescribePortfolio

Creates a value of DescribePortfolio with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

desId - The identifier of the portfolio for which to retrieve information.

data DescribePortfolio

See: describePortfolio smart constructor.

Request Lenses

desAcceptLanguage :: Lens' DescribePortfolio (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

desId :: Lens' DescribePortfolio Text

The identifier of the portfolio for which to retrieve information.

Destructuring the Response

describePortfolioResponse

Arguments

:: Int
dprsResponseStatus

-> DescribePortfolioResponse

Creates a value of DescribePortfolioResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsPortfolioDetail - Detailed portfolio information.

dprsTagOptions - TagOptions associated with the portfolio.

dprsTags - Tags associated with the portfolio.

dprsResponseStatus - -- | The response status code.

data DescribePortfolioResponse

See: describePortfolioResponse smart constructor.

Response Lenses

dprsPortfolioDetail :: Lens' DescribePortfolioResponse (Maybe PortfolioDetail)

Detailed portfolio information.

dprsTagOptions :: Lens' DescribePortfolioResponse [TagOptionDetail]

TagOptions associated with the portfolio.

dprsTags :: Lens' DescribePortfolioResponse [Tag]

Tags associated with the portfolio.

dprsResponseStatus :: Lens' DescribePortfolioResponse Int

- | The response status code.