Network.AWS.ServiceCatalog.DescribeProductAsAdmin

Description

Retrieves information about a specified product, run with administrator access.

Creating a Request

describeProductAsAdmin

Arguments

:: Text
dpaaId

-> DescribeProductAsAdmin

Creates a value of DescribeProductAsAdmin with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpaaAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dpaaId - The identifier of the product for which to retrieve information.

data DescribeProductAsAdmin

See: describeProductAsAdmin smart constructor.

Request Lenses

dpaaAcceptLanguage :: Lens' DescribeProductAsAdmin (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dpaaId :: Lens' DescribeProductAsAdmin Text

The identifier of the product for which to retrieve information.

Destructuring the Response

describeProductAsAdminResponse

Arguments

:: Int
dpaarsResponseStatus

-> DescribeProductAsAdminResponse

Creates a value of DescribeProductAsAdminResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpaarsProductViewDetail - Detailed product view information.

dpaarsTagOptions - List of TagOptions associated with the product.

dpaarsProvisioningArtifactSummaries - A list of provisioning artifact summaries for the product.

dpaarsTags - Tags associated with the product.

dpaarsResponseStatus - -- | The response status code.

data DescribeProductAsAdminResponse

See: describeProductAsAdminResponse smart constructor.

Response Lenses

dpaarsProductViewDetail :: Lens' DescribeProductAsAdminResponse (Maybe ProductViewDetail)

Detailed product view information.

dpaarsTagOptions :: Lens' DescribeProductAsAdminResponse [TagOptionDetail]

List of TagOptions associated with the product.

dpaarsProvisioningArtifactSummaries :: Lens' DescribeProductAsAdminResponse [ProvisioningArtifactSummary]

A list of provisioning artifact summaries for the product.

dpaarsTags :: Lens' DescribeProductAsAdminResponse [Tag]

Tags associated with the product.

dpaarsResponseStatus :: Lens' DescribeProductAsAdminResponse Int

- | The response status code.