Network.AWS.ServiceCatalog.DescribeProvisioningArtifact

Description

Retrieves detailed information about the specified provisioning artifact.

Creating a Request

describeProvisioningArtifact

Arguments

:: Text
dpaProvisioningArtifactId

-> Text
dpaProductId

-> DescribeProvisioningArtifact

Creates a value of DescribeProvisioningArtifact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpaVerbose - Enable a verbose level of details for the provisioning artifact.

dpaAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dpaProvisioningArtifactId - The identifier of the provisioning artifact. This is sometimes referred to as the product version.

dpaProductId - The product identifier.

data DescribeProvisioningArtifact

See: describeProvisioningArtifact smart constructor.

Request Lenses

dpaVerbose :: Lens' DescribeProvisioningArtifact (Maybe Bool)

Enable a verbose level of details for the provisioning artifact.

dpaAcceptLanguage :: Lens' DescribeProvisioningArtifact (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dpaProvisioningArtifactId :: Lens' DescribeProvisioningArtifact Text

The identifier of the provisioning artifact. This is sometimes referred to as the product version.

dpaProductId :: Lens' DescribeProvisioningArtifact Text

The product identifier.

Destructuring the Response

describeProvisioningArtifactResponse

Arguments

:: Int
dpaprsResponseStatus

-> DescribeProvisioningArtifactResponse

Creates a value of DescribeProvisioningArtifactResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpaprsStatus - The status of the current request.

dpaprsInfo - Additional information about the provisioning artifact.

dpaprsProvisioningArtifactDetail - Detailed provisioning artifact information.

dpaprsResponseStatus - -- | The response status code.

data DescribeProvisioningArtifactResponse

See: describeProvisioningArtifactResponse smart constructor.

Response Lenses

dpaprsStatus :: Lens' DescribeProvisioningArtifactResponse (Maybe RequestStatus)

The status of the current request.

dpaprsInfo :: Lens' DescribeProvisioningArtifactResponse (HashMap Text Text)

Additional information about the provisioning artifact.

dpaprsProvisioningArtifactDetail :: Lens' DescribeProvisioningArtifactResponse (Maybe ProvisioningArtifactDetail)

Detailed provisioning artifact information.

dpaprsResponseStatus :: Lens' DescribeProvisioningArtifactResponse Int

- | The response status code.