Network.AWS.ServiceCatalog.DescribeProvisioningParameters

Description

Provides information about parameters required to provision a specified product in a specified manner. Use this operation to obtain the list of ProvisioningArtifactParameters parameters available to call the ProvisionProduct operation for the specified product.

If the output contains a TagOption key with an empty list of values, there is a TagOption conflict for that key. The end user cannot take action to fix the conflict, and launch is not blocked. In subsequent calls to the ProvisionProduct operation, do not include conflicted TagOption keys as tags. Calls to ProvisionProduct with empty TagOption values cause the error "Parameter validation failed: Missing required parameter in Tags[N ]:Value ". Calls to ProvisionProduct with conflicted TagOption keys automatically tag the provisioned product with the conflicted keys with the value "sc-tagoption-conflict-portfolioId-productId ".

Creating a Request

describeProvisioningParameters

Arguments

:: Text
dppsProductId

-> Text
dppsProvisioningArtifactId

-> DescribeProvisioningParameters

Creates a value of DescribeProvisioningParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dppsAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dppsPathId - The identifier of the path for this product's provisioning. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

dppsProductId - The product identifier.

dppsProvisioningArtifactId - The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

data DescribeProvisioningParameters

See: describeProvisioningParameters smart constructor.

Request Lenses

dppsAcceptLanguage :: Lens' DescribeProvisioningParameters (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

dppsPathId :: Lens' DescribeProvisioningParameters (Maybe Text)

The identifier of the path for this product's provisioning. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

dppsProductId :: Lens' DescribeProvisioningParameters Text

The product identifier.

dppsProvisioningArtifactId :: Lens' DescribeProvisioningParameters Text

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

Destructuring the Response

describeProvisioningParametersResponse

Arguments

:: Int
dpprsResponseStatus

-> DescribeProvisioningParametersResponse

Creates a value of DescribeProvisioningParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpprsProvisioningArtifactParameters - The list of parameters used to successfully provision the product. Each parameter includes a list of allowable values and additional metadata about each parameter.

dpprsUsageInstructions - Any additional metadata specifically related to the provisioning of the product. For example, see the Version field of the CloudFormation template.

dpprsConstraintSummaries - The list of constraint summaries that apply to provisioning this product.

dpprsTagOptions - List of TagOptions associated with the provisioned provisioning parameters.

dpprsResponseStatus - -- | The response status code.

data DescribeProvisioningParametersResponse

See: describeProvisioningParametersResponse smart constructor.

Response Lenses

dpprsProvisioningArtifactParameters :: Lens' DescribeProvisioningParametersResponse [ProvisioningArtifactParameter]

The list of parameters used to successfully provision the product. Each parameter includes a list of allowable values and additional metadata about each parameter.

dpprsUsageInstructions :: Lens' DescribeProvisioningParametersResponse [UsageInstruction]

Any additional metadata specifically related to the provisioning of the product. For example, see the Version field of the CloudFormation template.

dpprsConstraintSummaries :: Lens' DescribeProvisioningParametersResponse [ConstraintSummary]

The list of constraint summaries that apply to provisioning this product.

dpprsTagOptions :: Lens' DescribeProvisioningParametersResponse [TagOptionSummary]

List of TagOptions associated with the provisioned provisioning parameters.

dpprsResponseStatus :: Lens' DescribeProvisioningParametersResponse Int

- | The response status code.