Network.AWS.ServiceCatalog.DescribeRecord

Description

Retrieves a paginated list of the full details of a specific request. Use this operation after calling a request operation (ProvisionProduct , TerminateProvisionedProduct , or UpdateProvisionedProduct ).

Creating a Request

describeRecord

Arguments

:: Text
drId

-> DescribeRecord

Creates a value of DescribeRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

drPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

drPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

drId - The record identifier of the ProvisionedProduct object for which to retrieve output information. This is the RecordDetail.RecordId obtained from the request operation's response.

data DescribeRecord

See: describeRecord smart constructor.

Request Lenses

drAcceptLanguage :: Lens' DescribeRecord (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

drPageToken :: Lens' DescribeRecord (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

drPageSize :: Lens' DescribeRecord (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

drId :: Lens' DescribeRecord Text

The record identifier of the ProvisionedProduct object for which to retrieve output information. This is the RecordDetail.RecordId obtained from the request operation's response.

Destructuring the Response

describeRecordResponse

Arguments

:: Int
drrsResponseStatus

-> DescribeRecordResponse

Creates a value of DescribeRecordResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrsRecordDetail - Detailed record information for the specified product.

drrsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

drrsRecordOutputs - A list of outputs for the specified Product object created as the result of a request. For example, a CloudFormation-backed product that creates an S3 bucket would have an output for the S3 bucket URL.

drrsResponseStatus - -- | The response status code.

data DescribeRecordResponse

See: describeRecordResponse smart constructor.

Response Lenses

drrsRecordDetail :: Lens' DescribeRecordResponse (Maybe RecordDetail)

Detailed record information for the specified product.

drrsNextPageToken :: Lens' DescribeRecordResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

drrsRecordOutputs :: Lens' DescribeRecordResponse [RecordOutput]

A list of outputs for the specified Product object created as the result of a request. For example, a CloudFormation-backed product that creates an S3 bucket would have an output for the S3 bucket URL.

drrsResponseStatus :: Lens' DescribeRecordResponse Int

- | The response status code.