Network.AWS.ServiceCatalog.DescribeTagOption

Description

Describes a TagOption.

Creating a Request

describeTagOption

Arguments

:: Text
dtoId

-> DescribeTagOption

Creates a value of DescribeTagOption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtoId - The identifier of the TagOption.

data DescribeTagOption

See: describeTagOption smart constructor.

Request Lenses

dtoId :: Lens' DescribeTagOption Text

The identifier of the TagOption.

Destructuring the Response

describeTagOptionResponse

Arguments

:: Int
dtorsResponseStatus

-> DescribeTagOptionResponse

Creates a value of DescribeTagOptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtorsTagOptionDetail - The resulting detailed TagOption information.

dtorsResponseStatus - -- | The response status code.

data DescribeTagOptionResponse

See: describeTagOptionResponse smart constructor.

Response Lenses

dtorsTagOptionDetail :: Lens' DescribeTagOptionResponse (Maybe TagOptionDetail)

The resulting detailed TagOption information.

dtorsResponseStatus :: Lens' DescribeTagOptionResponse Int

- | The response status code.