Network.AWS.ServiceCatalog.DisassociateTagOptionFromResource

Description

Disassociates a TagOption from a resource.

Creating a Request

disassociateTagOptionFromResource

Arguments

:: Text
dtofrResourceId

-> Text
dtofrTagOptionId

-> DisassociateTagOptionFromResource

Creates a value of DisassociateTagOptionFromResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtofrResourceId - Identifier of the resource from which to disassociate the TagOption.

dtofrTagOptionId - Identifier of the TagOption to disassociate from the resource.

data DisassociateTagOptionFromResource

See: disassociateTagOptionFromResource smart constructor.

Request Lenses

dtofrResourceId :: Lens' DisassociateTagOptionFromResource Text

Identifier of the resource from which to disassociate the TagOption.

dtofrTagOptionId :: Lens' DisassociateTagOptionFromResource Text

Identifier of the TagOption to disassociate from the resource.

Destructuring the Response

disassociateTagOptionFromResourceResponse

Arguments

:: Int
dtofrrsResponseStatus

-> DisassociateTagOptionFromResourceResponse

Creates a value of DisassociateTagOptionFromResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtofrrsResponseStatus - -- | The response status code.

data DisassociateTagOptionFromResourceResponse

See: disassociateTagOptionFromResourceResponse smart constructor.

Response Lenses

dtofrrsResponseStatus :: Lens' DisassociateTagOptionFromResourceResponse Int

- | The response status code.