Network.AWS.ServiceCatalog.ListConstraintsForPortfolio

Description

Retrieves detailed constraint information for the specified portfolio and product.

Creating a Request

listConstraintsForPortfolio

Arguments

:: Text
lcfpPortfolioId

-> ListConstraintsForPortfolio

Creates a value of ListConstraintsForPortfolio with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcfpAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lcfpPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lcfpPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

lcfpProductId - The product identifier.

lcfpPortfolioId - The portfolio identifier.

data ListConstraintsForPortfolio

See: listConstraintsForPortfolio smart constructor.

Request Lenses

lcfpAcceptLanguage :: Lens' ListConstraintsForPortfolio (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lcfpPageToken :: Lens' ListConstraintsForPortfolio (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lcfpPageSize :: Lens' ListConstraintsForPortfolio (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

lcfpProductId :: Lens' ListConstraintsForPortfolio (Maybe Text)

The product identifier.

lcfpPortfolioId :: Lens' ListConstraintsForPortfolio Text

The portfolio identifier.

Destructuring the Response

listConstraintsForPortfolioResponse

Arguments

:: Int
lcfprsResponseStatus

-> ListConstraintsForPortfolioResponse

Creates a value of ListConstraintsForPortfolioResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcfprsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lcfprsConstraintDetails - List of detailed constraint information objects.

lcfprsResponseStatus - -- | The response status code.

data ListConstraintsForPortfolioResponse

See: listConstraintsForPortfolioResponse smart constructor.

Response Lenses

lcfprsNextPageToken :: Lens' ListConstraintsForPortfolioResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lcfprsConstraintDetails :: Lens' ListConstraintsForPortfolioResponse [ConstraintDetail]

List of detailed constraint information objects.

lcfprsResponseStatus :: Lens' ListConstraintsForPortfolioResponse Int

- | The response status code.