Network.AWS.ServiceCatalog.ListPortfolioAccess

Description

Lists the account IDs that have been authorized sharing of the specified portfolio.

Creating a Request

listPortfolioAccess

Arguments

:: Text
lPortfolioId

-> ListPortfolioAccess

Creates a value of ListPortfolioAccess with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lPortfolioId - The portfolio identifier.

data ListPortfolioAccess

See: listPortfolioAccess smart constructor.

Request Lenses

lAcceptLanguage :: Lens' ListPortfolioAccess (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lPortfolioId :: Lens' ListPortfolioAccess Text

The portfolio identifier.

Destructuring the Response

listPortfolioAccessResponse

Arguments

:: Int
lparsResponseStatus

-> ListPortfolioAccessResponse

Creates a value of ListPortfolioAccessResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lparsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lparsAccountIds - List of account IDs associated with access to the portfolio.

lparsResponseStatus - -- | The response status code.

data ListPortfolioAccessResponse

See: listPortfolioAccessResponse smart constructor.

Response Lenses

lparsNextPageToken :: Lens' ListPortfolioAccessResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lparsAccountIds :: Lens' ListPortfolioAccessResponse [Text]

List of account IDs associated with access to the portfolio.

lparsResponseStatus :: Lens' ListPortfolioAccessResponse Int

- | The response status code.