Network.AWS.ServiceCatalog.ListPrincipalsForPortfolio

Description

Lists all principal ARNs associated with the specified portfolio.

Creating a Request

listPrincipalsForPortfolio

Arguments

:: Text
lpfpPortfolioId

-> ListPrincipalsForPortfolio

Creates a value of ListPrincipalsForPortfolio with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpfpAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lpfpPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lpfpPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

lpfpPortfolioId - The portfolio identifier.

data ListPrincipalsForPortfolio

See: listPrincipalsForPortfolio smart constructor.

Request Lenses

lpfpAcceptLanguage :: Lens' ListPrincipalsForPortfolio (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lpfpPageToken :: Lens' ListPrincipalsForPortfolio (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lpfpPageSize :: Lens' ListPrincipalsForPortfolio (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

lpfpPortfolioId :: Lens' ListPrincipalsForPortfolio Text

The portfolio identifier.

Destructuring the Response

listPrincipalsForPortfolioResponse

Arguments

:: Int
lisrsResponseStatus

-> ListPrincipalsForPortfolioResponse

Creates a value of ListPrincipalsForPortfolioResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lisrsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lisrsPrincipals - The IAM principals (users or roles) associated with the portfolio.

lisrsResponseStatus - -- | The response status code.

data ListPrincipalsForPortfolioResponse

See: listPrincipalsForPortfolioResponse smart constructor.

Response Lenses

lisrsNextPageToken :: Lens' ListPrincipalsForPortfolioResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lisrsPrincipals :: Lens' ListPrincipalsForPortfolioResponse [Principal]

The IAM principals (users or roles) associated with the portfolio.

lisrsResponseStatus :: Lens' ListPrincipalsForPortfolioResponse Int

- | The response status code.