Network.AWS.ServiceCatalog.ListRecordHistory

Description

Returns a paginated list of all performed requests, in the form of RecordDetails objects that are filtered as specified.

Creating a Request

listRecordHistory :: ListRecordHistory

Creates a value of ListRecordHistory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrhSearchFilter - The filter to limit search results.

lrhAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lrhAccessLevelFilter - The access level for obtaining results. If left unspecified, User level access is used.

lrhPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lrhPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

data ListRecordHistory

See: listRecordHistory smart constructor.

Request Lenses

lrhSearchFilter :: Lens' ListRecordHistory (Maybe ListRecordHistorySearchFilter)

The filter to limit search results.

lrhAcceptLanguage :: Lens' ListRecordHistory (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

lrhAccessLevelFilter :: Lens' ListRecordHistory (Maybe AccessLevelFilter)

The access level for obtaining results. If left unspecified, User level access is used.

lrhPageToken :: Lens' ListRecordHistory (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lrhPageSize :: Lens' ListRecordHistory (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

Destructuring the Response

listRecordHistoryResponse

Arguments

:: Int
lrhrsResponseStatus

-> ListRecordHistoryResponse

Creates a value of ListRecordHistoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrhrsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lrhrsRecordDetails - A list of record detail objects, listed in reverse chronological order.

lrhrsResponseStatus - -- | The response status code.

data ListRecordHistoryResponse

See: listRecordHistoryResponse smart constructor.

Response Lenses

lrhrsNextPageToken :: Lens' ListRecordHistoryResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

lrhrsRecordDetails :: Lens' ListRecordHistoryResponse [RecordDetail]

A list of record detail objects, listed in reverse chronological order.

lrhrsResponseStatus :: Lens' ListRecordHistoryResponse Int

- | The response status code.