Network.AWS.ServiceCatalog.ListResourcesForTagOption

Description

Lists resources associated with a TagOption.

Creating a Request

listResourcesForTagOption

Arguments

:: Text
lrftoTagOptionId

-> ListResourcesForTagOption

Creates a value of ListResourcesForTagOption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrftoResourceType - Resource type.

lrftoPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lrftoPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

lrftoTagOptionId - Identifier of the TagOption.

data ListResourcesForTagOption

See: listResourcesForTagOption smart constructor.

Request Lenses

lrftoResourceType :: Lens' ListResourcesForTagOption (Maybe Text)

Resource type.

lrftoPageToken :: Lens' ListResourcesForTagOption (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lrftoPageSize :: Lens' ListResourcesForTagOption (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

lrftoTagOptionId :: Lens' ListResourcesForTagOption Text

Identifier of the TagOption.

Destructuring the Response

listResourcesForTagOptionResponse

Arguments

:: Int
lrftorsResponseStatus

-> ListResourcesForTagOptionResponse

Creates a value of ListResourcesForTagOptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrftorsResourceDetails - The resulting detailed resource information.

lrftorsPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lrftorsResponseStatus - -- | The response status code.

data ListResourcesForTagOptionResponse

See: listResourcesForTagOptionResponse smart constructor.

Response Lenses

lrftorsResourceDetails :: Lens' ListResourcesForTagOptionResponse [ResourceDetail]

The resulting detailed resource information.

lrftorsPageToken :: Lens' ListResourcesForTagOptionResponse (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

lrftorsResponseStatus :: Lens' ListResourcesForTagOptionResponse Int

- | The response status code.