Network.AWS.ServiceCatalog.ListTagOptions

Description

Lists detailed TagOptions information.

Creating a Request

listTagOptions :: ListTagOptions

Creates a value of ListTagOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltoFilters - The list of filters with which to limit search results. If no search filters are specified, the output is all TagOptions.

ltoPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

ltoPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

data ListTagOptions

See: listTagOptions smart constructor.

Request Lenses

ltoFilters :: Lens' ListTagOptions (Maybe ListTagOptionsFilters)

The list of filters with which to limit search results. If no search filters are specified, the output is all TagOptions.

ltoPageToken :: Lens' ListTagOptions (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

ltoPageSize :: Lens' ListTagOptions (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

Destructuring the Response

listTagOptionsResponse

Arguments

:: Int
ltorsResponseStatus

-> ListTagOptionsResponse

Creates a value of ListTagOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltorsPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

ltorsTagOptionDetails - The resulting detailed TagOption information.

ltorsResponseStatus - -- | The response status code.

data ListTagOptionsResponse

See: listTagOptionsResponse smart constructor.

Response Lenses

ltorsPageToken :: Lens' ListTagOptionsResponse (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

ltorsTagOptionDetails :: Lens' ListTagOptionsResponse [TagOptionDetail]

The resulting detailed TagOption information.

ltorsResponseStatus :: Lens' ListTagOptionsResponse Int

- | The response status code.