Network.AWS.ServiceCatalog.ProvisionProduct

Description

Requests a provision of a specified product. A provisioned product is a resourced instance for a product. For example, provisioning a CloudFormation-template-backed product results in launching a CloudFormation stack and all the underlying resources that come with it.

You can check the status of this request using the DescribeRecord operation. The error "Parameter validation failed: Missing required parameter in Tags[N ]:Value " indicates that your request contains a tag which has a tag key but no corresponding tag value (value is empty or null). Your call may have included values returned from a DescribeProvisioningParameters call that resulted in a TagOption key with an empty list. This happens when TagOption keys are in conflict. For more information, see DescribeProvisioningParameters .

Creating a Request

provisionProduct

Arguments

:: Text
ppProductId

-> Text
ppProvisioningArtifactId

-> Text
ppProvisionedProductName

-> Text
ppProvisionToken

-> ProvisionProduct

Creates a value of ProvisionProduct with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppNotificationARNs - Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

ppAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

ppPathId - The identifier of the path for this product's provisioning. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

ppProvisioningParameters - Parameters specified by the administrator that are required for provisioning the product.

ppTags - A list of tags to use as provisioning options.

ppProductId - The product identifier.

ppProvisioningArtifactId - The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

ppProvisionedProductName - A user-friendly name to identify the ProvisionedProduct object. This value must be unique for the AWS account and cannot be updated after the product is provisioned.

ppProvisionToken - An idempotency token that uniquely identifies the provisioning request.

data ProvisionProduct

See: provisionProduct smart constructor.

Request Lenses

ppNotificationARNs :: Lens' ProvisionProduct [Text]

Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

ppAcceptLanguage :: Lens' ProvisionProduct (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

ppPathId :: Lens' ProvisionProduct (Maybe Text)

The identifier of the path for this product's provisioning. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

ppProvisioningParameters :: Lens' ProvisionProduct [ProvisioningParameter]

Parameters specified by the administrator that are required for provisioning the product.

ppTags :: Lens' ProvisionProduct [Tag]

A list of tags to use as provisioning options.

ppProductId :: Lens' ProvisionProduct Text

The product identifier.

ppProvisioningArtifactId :: Lens' ProvisionProduct Text

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

ppProvisionedProductName :: Lens' ProvisionProduct Text

A user-friendly name to identify the ProvisionedProduct object. This value must be unique for the AWS account and cannot be updated after the product is provisioned.

ppProvisionToken :: Lens' ProvisionProduct Text

An idempotency token that uniquely identifies the provisioning request.

Destructuring the Response

provisionProductResponse

Arguments

:: Int
pprsResponseStatus

-> ProvisionProductResponse

Creates a value of ProvisionProductResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pprsRecordDetail - The detailed result of the ProvisionProduct request, containing the inputs made to that request, the current state of the request, a pointer to the ProvisionedProduct object of the request, and a list of any errors that the request encountered.

pprsResponseStatus - -- | The response status code.

data ProvisionProductResponse

See: provisionProductResponse smart constructor.

Response Lenses

pprsRecordDetail :: Lens' ProvisionProductResponse (Maybe RecordDetail)

The detailed result of the ProvisionProduct request, containing the inputs made to that request, the current state of the request, a pointer to the ProvisionedProduct object of the request, and a list of any errors that the request encountered.

pprsResponseStatus :: Lens' ProvisionProductResponse Int

- | The response status code.