Network.AWS.ServiceCatalog.SearchProducts

Description

Returns a paginated list all of the Products objects to which the caller has access.

The output of this operation can be used as input for other operations, such as DescribeProductView .

Creating a Request

searchProducts :: SearchProducts

Creates a value of SearchProducts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spFilters - The list of filters with which to limit search results. If no search filters are specified, the output is all the products to which the calling user has access.

spSortOrder - The sort order specifier. If no value is specified, results are not sorted.

spAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

spPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

spPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

spSortBy - The sort field specifier. If no value is specified, results are not sorted.

data SearchProducts

See: searchProducts smart constructor.

Request Lenses

spFilters :: Lens' SearchProducts (HashMap ProductViewFilterBy [Text])

The list of filters with which to limit search results. If no search filters are specified, the output is all the products to which the calling user has access.

spSortOrder :: Lens' SearchProducts (Maybe SortOrder)

The sort order specifier. If no value is specified, results are not sorted.

spAcceptLanguage :: Lens' SearchProducts (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

spPageToken :: Lens' SearchProducts (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

spPageSize :: Lens' SearchProducts (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

spSortBy :: Lens' SearchProducts (Maybe ProductViewSortBy)

The sort field specifier. If no value is specified, results are not sorted.

Destructuring the Response

searchProductsResponse

Arguments

:: Int
sprsResponseStatus

-> SearchProductsResponse

Creates a value of SearchProductsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sprsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

sprsProductViewAggregations - A list of the product view aggregation value objects.

sprsProductViewSummaries - A list of the product view summary objects.

sprsResponseStatus - -- | The response status code.

data SearchProductsResponse

See: searchProductsResponse smart constructor.

Response Lenses

sprsNextPageToken :: Lens' SearchProductsResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

sprsProductViewAggregations :: Lens' SearchProductsResponse (HashMap Text [ProductViewAggregationValue])

A list of the product view aggregation value objects.

sprsProductViewSummaries :: Lens' SearchProductsResponse [ProductViewSummary]

A list of the product view summary objects.

sprsResponseStatus :: Lens' SearchProductsResponse Int

- | The response status code.