Network.AWS.ServiceCatalog.SearchProductsAsAdmin

Description

Retrieves summary and status information about all products created within the caller's account. If a portfolio ID is provided, this operation retrieves information for only those products that are associated with the specified portfolio.

Creating a Request

searchProductsAsAdmin :: SearchProductsAsAdmin

Creates a value of SearchProductsAsAdmin with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spaaPortfolioId - The portfolio identifier.

spaaFilters - The list of filters with which to limit search results. If no search filters are specified, the output is all the products to which the administrator has access.

spaaSortOrder - The sort order specifier. If no value is specified, results are not sorted.

spaaAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

spaaPageToken - The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

spaaPageSize - The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

spaaProductSource - Access level of the source of the product.

spaaSortBy - The sort field specifier. If no value is specified, results are not sorted.

data SearchProductsAsAdmin

See: searchProductsAsAdmin smart constructor.

Request Lenses

spaaPortfolioId :: Lens' SearchProductsAsAdmin (Maybe Text)

The portfolio identifier.

spaaFilters :: Lens' SearchProductsAsAdmin (HashMap ProductViewFilterBy [Text])

The list of filters with which to limit search results. If no search filters are specified, the output is all the products to which the administrator has access.

spaaSortOrder :: Lens' SearchProductsAsAdmin (Maybe SortOrder)

The sort order specifier. If no value is specified, results are not sorted.

spaaAcceptLanguage :: Lens' SearchProductsAsAdmin (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

spaaPageToken :: Lens' SearchProductsAsAdmin (Maybe Text)

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize .

spaaPageSize :: Lens' SearchProductsAsAdmin (Maybe Natural)

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize , the value of NextPageToken in the response is non-null.

spaaProductSource :: Lens' SearchProductsAsAdmin (Maybe ProductSource)

Access level of the source of the product.

spaaSortBy :: Lens' SearchProductsAsAdmin (Maybe ProductViewSortBy)

The sort field specifier. If no value is specified, results are not sorted.

Destructuring the Response

searchProductsAsAdminResponse

Arguments

:: Int
spaarsResponseStatus

-> SearchProductsAsAdminResponse

Creates a value of SearchProductsAsAdminResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spaarsNextPageToken - The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

spaarsProductViewDetails - List of detailed product view information objects.

spaarsResponseStatus - -- | The response status code.

data SearchProductsAsAdminResponse

See: searchProductsAsAdminResponse smart constructor.

Response Lenses

spaarsNextPageToken :: Lens' SearchProductsAsAdminResponse (Maybe Text)

The page token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

spaarsProductViewDetails :: Lens' SearchProductsAsAdminResponse [ProductViewDetail]

List of detailed product view information objects.

spaarsResponseStatus :: Lens' SearchProductsAsAdminResponse Int

- | The response status code.