Network.AWS.ServiceCatalog.Types

Description

Service Configuration

serviceCatalog :: Service

API version 2015-12-10 of the Amazon Service Catalog SDK configuration.

Errors

_InvalidParametersException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters provided to the operation are invalid.

_DuplicateResourceException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is a duplicate.

_TagOptionNotMigratedException :: AsError a => Getting (First ServiceError) a ServiceError

An operation requiring TagOptions failed because the TagOptions migration process has not been performed for this account. Please use the AWS console to perform the migration process before retrying the operation.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

_InvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

An attempt was made to modify a resource that is in an invalid state. Inspect the resource you are using for this operation to ensure that all resource states are valid before retrying the operation.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The current limits of the service would have been exceeded by this operation. Reduce the resource use or increase the service limits and retry the operation.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The operation was requested against a resource that is currently in use. Free the resource from use and retry the operation.

AccessLevelFilterKey

data AccessLevelFilterKey

Constructors

Account

Role

User

PrincipalType

data PrincipalType

Constructors

IAM

ProductSource

data ProductSource

Constructors

PSAccount

ProductType

data ProductType

Constructors

CloudFormationTemplate

Marketplace

ProductViewFilterBy

data ProductViewFilterBy

Constructors

FullTextSearch

Owner

ProductType

SourceProductId

ProductViewSortBy

data ProductViewSortBy

Constructors

CreationDate

Title

VersionCount

ProvisionedProductStatus

data ProvisionedProductStatus

Constructors

PPSAvailable

PPSError'

PPSTainted

PPSUnderChange

ProvisioningArtifactType

data ProvisioningArtifactType

Constructors

PATCloudFormationTemplate

PATMarketplaceAMI

PATMarketplaceCar

RecordStatus

data RecordStatus

Constructors

RSCreated

RSFailed

RSInProgress

RSInProgressInError

RSSucceeded

RequestStatus

data RequestStatus

Constructors

Available

Creating

Failed

SortOrder

data SortOrder

Constructors

Ascending

Descending

AccessLevelFilter

data AccessLevelFilter

The access level to limit results.

See: accessLevelFilter smart constructor.

accessLevelFilter :: AccessLevelFilter

Creates a value of AccessLevelFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

alfValue - Specifies the user to which the access level applies. A value of Self is currently supported.

alfKey - Specifies the access level. Account allows results at the account level. Role allows results based on the federated role of the specified user. User allows results limited to the specified user.

alfValue :: Lens' AccessLevelFilter (Maybe Text)

Specifies the user to which the access level applies. A value of Self is currently supported.

alfKey :: Lens' AccessLevelFilter (Maybe AccessLevelFilterKey)

Specifies the access level. Account allows results at the account level. Role allows results based on the federated role of the specified user. User allows results limited to the specified user.

ConstraintDetail

data ConstraintDetail

Detailed constraint information.

See: constraintDetail smart constructor.

constraintDetail :: ConstraintDetail

Creates a value of ConstraintDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdConstraintId - The identifier of the constraint.

cdOwner - The owner of the constraint.

cdType - The type of the constraint.

cdDescription - The text description of the constraint.

cdConstraintId :: Lens' ConstraintDetail (Maybe Text)

The identifier of the constraint.

cdOwner :: Lens' ConstraintDetail (Maybe Text)

The owner of the constraint.

cdType :: Lens' ConstraintDetail (Maybe Text)

The type of the constraint.

cdDescription :: Lens' ConstraintDetail (Maybe Text)

The text description of the constraint.

ConstraintSummary

data ConstraintSummary

An administrator-specified constraint to apply when provisioning a product.

See: constraintSummary smart constructor.

constraintSummary :: ConstraintSummary

Creates a value of ConstraintSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csType - The type of the constraint.

csDescription - The text description of the constraint.

csType :: Lens' ConstraintSummary (Maybe Text)

The type of the constraint.

csDescription :: Lens' ConstraintSummary (Maybe Text)

The text description of the constraint.

LaunchPathSummary

data LaunchPathSummary

Summary information about a path for a user to have access to a specified product.

See: launchPathSummary smart constructor.

launchPathSummary :: LaunchPathSummary

Creates a value of LaunchPathSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpsConstraintSummaries - List of constraints on the portfolio-product relationship.

lpsName - Corresponds to the name of the portfolio to which the user was assigned.

lpsId - The unique identifier of the product path.

lpsTags - List of tags used by this launch path.

lpsConstraintSummaries :: Lens' LaunchPathSummary [ConstraintSummary]

List of constraints on the portfolio-product relationship.

lpsName :: Lens' LaunchPathSummary (Maybe Text)

Corresponds to the name of the portfolio to which the user was assigned.

lpsId :: Lens' LaunchPathSummary (Maybe Text)

The unique identifier of the product path.

lpsTags :: Lens' LaunchPathSummary [Tag]

List of tags used by this launch path.

ListRecordHistorySearchFilter

data ListRecordHistorySearchFilter

The search filter to limit results when listing request history records.

See: listRecordHistorySearchFilter smart constructor.

listRecordHistorySearchFilter :: ListRecordHistorySearchFilter

Creates a value of ListRecordHistorySearchFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrhsfValue - The filter value for Key .

lrhsfKey - The filter key.

lrhsfValue :: Lens' ListRecordHistorySearchFilter (Maybe Text)

The filter value for Key .

lrhsfKey :: Lens' ListRecordHistorySearchFilter (Maybe Text)

The filter key.

ListTagOptionsFilters

data ListTagOptionsFilters

The ListTagOptions filters.

See: listTagOptionsFilters smart constructor.

listTagOptionsFilters :: ListTagOptionsFilters

Creates a value of ListTagOptionsFilters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltofValue - The ListTagOptionsFilters value.

ltofActive - The ListTagOptionsFilters active state.

ltofKey - The ListTagOptionsFilters key.

ltofValue :: Lens' ListTagOptionsFilters (Maybe Text)

The ListTagOptionsFilters value.

ltofActive :: Lens' ListTagOptionsFilters (Maybe Bool)

The ListTagOptionsFilters active state.

ltofKey :: Lens' ListTagOptionsFilters (Maybe Text)

The ListTagOptionsFilters key.

ParameterConstraints

data ParameterConstraints

The constraints that the administrator has put on the parameter.

See: parameterConstraints smart constructor.

parameterConstraints :: ParameterConstraints

Creates a value of ParameterConstraints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pcAllowedValues - The values that the administrator has allowed for the parameter.

pcAllowedValues :: Lens' ParameterConstraints [Text]

The values that the administrator has allowed for the parameter.

PortfolioDetail

data PortfolioDetail

Detailed portfolio information.

See: portfolioDetail smart constructor.

portfolioDetail :: PortfolioDetail

Creates a value of PortfolioDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pdARN - The ARN assigned to the portfolio.

pdCreatedTime - The UTC timestamp of the creation time.

pdId - The identifier for the portfolio.

pdDisplayName - The name to use for display purposes.

pdDescription - The text description of the portfolio.

pdProviderName - The name of the portfolio provider.

pdARN :: Lens' PortfolioDetail (Maybe Text)

The ARN assigned to the portfolio.

pdCreatedTime :: Lens' PortfolioDetail (Maybe UTCTime)

The UTC timestamp of the creation time.

pdId :: Lens' PortfolioDetail (Maybe Text)

The identifier for the portfolio.

pdDisplayName :: Lens' PortfolioDetail (Maybe Text)

The name to use for display purposes.

pdDescription :: Lens' PortfolioDetail (Maybe Text)

The text description of the portfolio.

pdProviderName :: Lens' PortfolioDetail (Maybe Text)

The name of the portfolio provider.

Principal

data Principal

A principal's ARN and type.

See: principal smart constructor.

principal :: Principal

Creates a value of Principal with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pPrincipalType - The principal type. Must be IAM

pPrincipalARN - The ARN representing the principal (IAM user, role, or group).

pPrincipalType :: Lens' Principal (Maybe PrincipalType)

The principal type. Must be IAM

pPrincipalARN :: Lens' Principal (Maybe Text)

The ARN representing the principal (IAM user, role, or group).

ProductViewAggregationValue

data ProductViewAggregationValue

A single product view aggregation value/count pair, containing metadata about each product to which the calling user has access.

See: productViewAggregationValue smart constructor.

productViewAggregationValue :: ProductViewAggregationValue

Creates a value of ProductViewAggregationValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pvavValue - The value of the product view aggregation.

pvavApproximateCount - An approximate count of the products that match the value.

pvavValue :: Lens' ProductViewAggregationValue (Maybe Text)

The value of the product view aggregation.

pvavApproximateCount :: Lens' ProductViewAggregationValue (Maybe Int)

An approximate count of the products that match the value.

ProductViewDetail

data ProductViewDetail

Detailed product view information.

See: productViewDetail smart constructor.

productViewDetail :: ProductViewDetail

Creates a value of ProductViewDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pvdStatus - Current status of the product. AVAILABLE - Product is available for use. CREATING - Creation of product started, not ready for use. FAILED - Action on product failed.

pvdProductViewSummary - The summary metadata about the specified product view.

pvdCreatedTime - The UTC timestamp of the creation time.

pvdProductARN - The ARN associated with the product.

pvdStatus :: Lens' ProductViewDetail (Maybe RequestStatus)

Current status of the product. AVAILABLE - Product is available for use. CREATING - Creation of product started, not ready for use. FAILED - Action on product failed.

pvdProductViewSummary :: Lens' ProductViewDetail (Maybe ProductViewSummary)

The summary metadata about the specified product view.

pvdCreatedTime :: Lens' ProductViewDetail (Maybe UTCTime)

The UTC timestamp of the creation time.

pvdProductARN :: Lens' ProductViewDetail (Maybe Text)

The ARN associated with the product.

ProductViewSummary

data ProductViewSummary

The summary metadata about the specified product.

See: productViewSummary smart constructor.

productViewSummary :: ProductViewSummary

Creates a value of ProductViewSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pvsOwner - The owner of the product. Contact the product administrator for the significance of this value.

pvsSupportURL - The URL information to obtain support for this Product.

pvsShortDescription - Short description of the product.

pvsHasDefaultPath - A value of false indicates that the product does not have a default path, while a value of true indicates that it does. If it's false, call ListLaunchPaths to disambiguate between paths. If true, ListLaunchPaths is not required, and the output of the ProductViewSummary operation can be used directly with DescribeProvisioningParameters .

pvsDistributor - The distributor of the product. Contact the product administrator for the significance of this value.

pvsName - The name of the product.

pvsId - The product view identifier.

pvsType - The product type. Contact the product administrator for the significance of this value. If this value is MARKETPLACE , the product was created by AWS Marketplace.

pvsSupportEmail - The email contact information to obtain support for this Product.

pvsProductId - The product identifier.

pvsSupportDescription - The description of the support for this Product.

pvsOwner :: Lens' ProductViewSummary (Maybe Text)

The owner of the product. Contact the product administrator for the significance of this value.

pvsSupportURL :: Lens' ProductViewSummary (Maybe Text)

The URL information to obtain support for this Product.

pvsShortDescription :: Lens' ProductViewSummary (Maybe Text)

Short description of the product.

pvsHasDefaultPath :: Lens' ProductViewSummary (Maybe Bool)

A value of false indicates that the product does not have a default path, while a value of true indicates that it does. If it's false, call ListLaunchPaths to disambiguate between paths. If true, ListLaunchPaths is not required, and the output of the ProductViewSummary operation can be used directly with DescribeProvisioningParameters .

pvsDistributor :: Lens' ProductViewSummary (Maybe Text)

The distributor of the product. Contact the product administrator for the significance of this value.

pvsName :: Lens' ProductViewSummary (Maybe Text)

The name of the product.

pvsId :: Lens' ProductViewSummary (Maybe Text)

The product view identifier.

pvsType :: Lens' ProductViewSummary (Maybe ProductType)

The product type. Contact the product administrator for the significance of this value. If this value is MARKETPLACE , the product was created by AWS Marketplace.

pvsSupportEmail :: Lens' ProductViewSummary (Maybe Text)

The email contact information to obtain support for this Product.

pvsProductId :: Lens' ProductViewSummary (Maybe Text)

The product identifier.

pvsSupportDescription :: Lens' ProductViewSummary (Maybe Text)

The description of the support for this Product.

ProvisionedProductDetail

data ProvisionedProductDetail

Detailed information about a ProvisionedProduct object.

See: provisionedProductDetail smart constructor.

provisionedProductDetail :: ProvisionedProductDetail

Creates a value of ProvisionedProductDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppdIdempotencyToken - A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

ppdStatus - The current status of the ProvisionedProduct. AVAILABLE - Stable state, ready to perform any operation. The most recent action request succeeded and completed. UNDER_CHANGE - Transitive state, operations performed may or may not have valid results. Wait for an AVAILABLE status before performing operations. TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version. ERROR - Something unexpected happened such that the provisioned product exists but the stack is not running. For example, CloudFormation received an invalid parameter value and could not launch the stack.

ppdARN - The ARN associated with the ProvisionedProduct object.

ppdCreatedTime - The UTC timestamp of the creation time.

ppdStatusMessage - The current status message of the ProvisionedProduct.

ppdName - The user-friendly name of the ProvisionedProduct object.

ppdLastRecordId - The record identifier of the last request performed on this ProvisionedProduct object.

ppdId - The identifier of the ProvisionedProduct object.

ppdType - The type of the ProvisionedProduct object.

ppdIdempotencyToken :: Lens' ProvisionedProductDetail (Maybe Text)

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

ppdStatus :: Lens' ProvisionedProductDetail (Maybe ProvisionedProductStatus)

The current status of the ProvisionedProduct. AVAILABLE - Stable state, ready to perform any operation. The most recent action request succeeded and completed. UNDER_CHANGE - Transitive state, operations performed may or may not have valid results. Wait for an AVAILABLE status before performing operations. TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version. ERROR - Something unexpected happened such that the provisioned product exists but the stack is not running. For example, CloudFormation received an invalid parameter value and could not launch the stack.

ppdARN :: Lens' ProvisionedProductDetail (Maybe Text)

The ARN associated with the ProvisionedProduct object.

ppdCreatedTime :: Lens' ProvisionedProductDetail (Maybe UTCTime)

The UTC timestamp of the creation time.

ppdStatusMessage :: Lens' ProvisionedProductDetail (Maybe Text)

The current status message of the ProvisionedProduct.

ppdName :: Lens' ProvisionedProductDetail (Maybe Text)

The user-friendly name of the ProvisionedProduct object.

ppdLastRecordId :: Lens' ProvisionedProductDetail (Maybe Text)

The record identifier of the last request performed on this ProvisionedProduct object.

ppdId :: Lens' ProvisionedProductDetail (Maybe Text)

The identifier of the ProvisionedProduct object.

ppdType :: Lens' ProvisionedProductDetail (Maybe Text)

The type of the ProvisionedProduct object.

ProvisioningArtifact

data ProvisioningArtifact

Contains information indicating the ways in which a product can be provisioned.

See: provisioningArtifact smart constructor.

provisioningArtifact :: ProvisioningArtifact

Creates a value of ProvisioningArtifact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paCreatedTime - The UTC timestamp of the creation time.

paName - The name of the artifact.

paId - The identifier for the artifact. This is sometimes referred to as the product version.

paDescription - The text description of the artifact.

paCreatedTime :: Lens' ProvisioningArtifact (Maybe UTCTime)

The UTC timestamp of the creation time.

paName :: Lens' ProvisioningArtifact (Maybe Text)

The name of the artifact.

paId :: Lens' ProvisioningArtifact (Maybe Text)

The identifier for the artifact. This is sometimes referred to as the product version.

paDescription :: Lens' ProvisioningArtifact (Maybe Text)

The text description of the artifact.

ProvisioningArtifactDetail

data ProvisioningArtifactDetail

Detailed provisioning artifact information.

See: provisioningArtifactDetail smart constructor.

provisioningArtifactDetail :: ProvisioningArtifactDetail

Creates a value of ProvisioningArtifactDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

padCreatedTime - The UTC timestamp of the creation time.

padName - The name assigned to the provisioning artifact.

padId - The identifier of the provisioning artifact. This is sometimes referred to as the product version.

padType - The type of the provisioning artifact. The following provisioning artifact types are used by AWS Marketplace products: MARKETPLACE_AMI - AMI products. MARKETPLACE_CAR - CAR (Cluster and AWS Resources) products.

padDescription - The text description of the provisioning artifact.

padCreatedTime :: Lens' ProvisioningArtifactDetail (Maybe UTCTime)

The UTC timestamp of the creation time.

padName :: Lens' ProvisioningArtifactDetail (Maybe Text)

The name assigned to the provisioning artifact.

padId :: Lens' ProvisioningArtifactDetail (Maybe Text)

The identifier of the provisioning artifact. This is sometimes referred to as the product version.

padType :: Lens' ProvisioningArtifactDetail (Maybe ProvisioningArtifactType)

The type of the provisioning artifact. The following provisioning artifact types are used by AWS Marketplace products: MARKETPLACE_AMI - AMI products. MARKETPLACE_CAR - CAR (Cluster and AWS Resources) products.

padDescription :: Lens' ProvisioningArtifactDetail (Maybe Text)

The text description of the provisioning artifact.

ProvisioningArtifactParameter

data ProvisioningArtifactParameter

A parameter used to successfully provision the product. This value includes a list of allowable values and additional metadata.

See: provisioningArtifactParameter smart constructor.

provisioningArtifactParameter :: ProvisioningArtifactParameter

Creates a value of ProvisioningArtifactParameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pIsNoEcho - If this value is true, the value for this parameter is obfuscated from view when the parameter is retrieved. This parameter is used to hide sensitive information.

pParameterKey - The parameter key.

pParameterType - The parameter type.

pParameterConstraints - The list of constraints that the administrator has put on the parameter.

pDefaultValue - The default value for this parameter.

pDescription - The text description of the parameter.

pIsNoEcho :: Lens' ProvisioningArtifactParameter (Maybe Bool)

If this value is true, the value for this parameter is obfuscated from view when the parameter is retrieved. This parameter is used to hide sensitive information.

pParameterKey :: Lens' ProvisioningArtifactParameter (Maybe Text)

The parameter key.

pParameterType :: Lens' ProvisioningArtifactParameter (Maybe Text)

The parameter type.

pParameterConstraints :: Lens' ProvisioningArtifactParameter (Maybe ParameterConstraints)

The list of constraints that the administrator has put on the parameter.

pDefaultValue :: Lens' ProvisioningArtifactParameter (Maybe Text)

The default value for this parameter.

pDescription :: Lens' ProvisioningArtifactParameter (Maybe Text)

The text description of the parameter.

ProvisioningArtifactProperties

data ProvisioningArtifactProperties

Provisioning artifact properties. For example request JSON, see CreateProvisioningArtifact .

See: provisioningArtifactProperties smart constructor.

provisioningArtifactProperties :: ProvisioningArtifactProperties

Creates a value of ProvisioningArtifactProperties with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

papName - The name assigned to the provisioning artifact properties.

papType - The type of the provisioning artifact properties. The following provisioning artifact property types are used by AWS Marketplace products: MARKETPLACE_AMI - AMI products. MARKETPLACE_CAR - CAR (Cluster and AWS Resources) products.

papDescription - The text description of the provisioning artifact properties.

papInfo - Additional information about the provisioning artifact properties. When using this element in a request, you must specify LoadTemplateFromURL . For more information, see CreateProvisioningArtifact .

papName :: Lens' ProvisioningArtifactProperties (Maybe Text)

The name assigned to the provisioning artifact properties.

papType :: Lens' ProvisioningArtifactProperties (Maybe ProvisioningArtifactType)

The type of the provisioning artifact properties. The following provisioning artifact property types are used by AWS Marketplace products: MARKETPLACE_AMI - AMI products. MARKETPLACE_CAR - CAR (Cluster and AWS Resources) products.

papDescription :: Lens' ProvisioningArtifactProperties (Maybe Text)

The text description of the provisioning artifact properties.

papInfo :: Lens' ProvisioningArtifactProperties (HashMap Text Text)

Additional information about the provisioning artifact properties. When using this element in a request, you must specify LoadTemplateFromURL . For more information, see CreateProvisioningArtifact .

ProvisioningArtifactSummary

data ProvisioningArtifactSummary

Stores summary information about a provisioning artifact.

See: provisioningArtifactSummary smart constructor.

provisioningArtifactSummary :: ProvisioningArtifactSummary

Creates a value of ProvisioningArtifactSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pasProvisioningArtifactMetadata - The provisioning artifact metadata. This data is used with products created by AWS Marketplace.

pasCreatedTime - The UTC timestamp of the creation time.

pasName - The name of the provisioning artifact.

pasId - The identifier of the provisioning artifact.

pasDescription - The description of the provisioning artifact.

pasProvisioningArtifactMetadata :: Lens' ProvisioningArtifactSummary (HashMap Text Text)

The provisioning artifact metadata. This data is used with products created by AWS Marketplace.

pasCreatedTime :: Lens' ProvisioningArtifactSummary (Maybe UTCTime)

The UTC timestamp of the creation time.

pasName :: Lens' ProvisioningArtifactSummary (Maybe Text)

The name of the provisioning artifact.

pasId :: Lens' ProvisioningArtifactSummary (Maybe Text)

The identifier of the provisioning artifact.

pasDescription :: Lens' ProvisioningArtifactSummary (Maybe Text)

The description of the provisioning artifact.

ProvisioningParameter

data ProvisioningParameter

The parameter key-value pairs used to provision a product.

See: provisioningParameter smart constructor.

provisioningParameter :: ProvisioningParameter

Creates a value of ProvisioningParameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppValue - The value to use for provisioning. Any constraints on this value can be found in ProvisioningArtifactParameter for Key .

ppKey - The ProvisioningArtifactParameter.ParameterKey parameter from DescribeProvisioningParameters .

ppValue :: Lens' ProvisioningParameter (Maybe Text)

The value to use for provisioning. Any constraints on this value can be found in ProvisioningArtifactParameter for Key .

ppKey :: Lens' ProvisioningParameter (Maybe Text)

The ProvisioningArtifactParameter.ParameterKey parameter from DescribeProvisioningParameters .

RecordDetail

data RecordDetail

The full details of a specific ProvisionedProduct object.

See: recordDetail smart constructor.

recordDetail :: RecordDetail

Creates a value of RecordDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdStatus - The status of the ProvisionedProduct object. CREATED - Request created but the operation has not yet started. IN_PROGRESS - The requested operation is in-progress. IN_PROGRESS_IN_ERROR - The provisioned product is under change but the requested operation failed and some remediation is occurring. For example, a rollback. SUCCEEDED - The requested operation has successfully completed. FAILED - The requested operation has completed but has failed. Investigate using the error messages returned.

rdRecordTags - List of tags associated with this record.

rdProvisionedProductName - The user-friendly name of the ProvisionedProduct object.

rdProvisioningArtifactId - The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

rdCreatedTime - The UTC timestamp of the creation time.

rdRecordType - The record type for this record.

rdRecordId - The identifier of the ProvisionedProduct object record.

rdProvisionedProductType - The type of the ProvisionedProduct object.

rdUpdatedTime - The time when the record for the ProvisionedProduct object was last updated.

rdPathId - The identifier of the path for this product's provisioning.

rdProvisionedProductId - The identifier of the ProvisionedProduct object.

rdRecordErrors - A list of errors that occurred while processing the request.

rdProductId - The product identifier.

rdStatus :: Lens' RecordDetail (Maybe RecordStatus)

The status of the ProvisionedProduct object. CREATED - Request created but the operation has not yet started. IN_PROGRESS - The requested operation is in-progress. IN_PROGRESS_IN_ERROR - The provisioned product is under change but the requested operation failed and some remediation is occurring. For example, a rollback. SUCCEEDED - The requested operation has successfully completed. FAILED - The requested operation has completed but has failed. Investigate using the error messages returned.

rdRecordTags :: Lens' RecordDetail [RecordTag]

List of tags associated with this record.

rdProvisionedProductName :: Lens' RecordDetail (Maybe Text)

The user-friendly name of the ProvisionedProduct object.

rdProvisioningArtifactId :: Lens' RecordDetail (Maybe Text)

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

rdCreatedTime :: Lens' RecordDetail (Maybe UTCTime)

The UTC timestamp of the creation time.

rdRecordType :: Lens' RecordDetail (Maybe Text)

The record type for this record.

rdRecordId :: Lens' RecordDetail (Maybe Text)

The identifier of the ProvisionedProduct object record.

rdProvisionedProductType :: Lens' RecordDetail (Maybe Text)

The type of the ProvisionedProduct object.

rdUpdatedTime :: Lens' RecordDetail (Maybe UTCTime)

The time when the record for the ProvisionedProduct object was last updated.

rdPathId :: Lens' RecordDetail (Maybe Text)

The identifier of the path for this product's provisioning.

rdProvisionedProductId :: Lens' RecordDetail (Maybe Text)

The identifier of the ProvisionedProduct object.

rdRecordErrors :: Lens' RecordDetail [RecordError]

A list of errors that occurred while processing the request.

rdProductId :: Lens' RecordDetail (Maybe Text)

The product identifier.

RecordError

data RecordError

The error code and description resulting from an operation.

See: recordError smart constructor.

recordError :: RecordError

Creates a value of RecordError with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

reCode - The numeric value of the error.

reDescription - The text description of the error.

reCode :: Lens' RecordError (Maybe Text)

The numeric value of the error.

reDescription :: Lens' RecordError (Maybe Text)

The text description of the error.

RecordOutput

data RecordOutput

An output for the specified Product object created as the result of a request. For example, a CloudFormation-backed product that creates an S3 bucket would have an output for the S3 bucket URL.

See: recordOutput smart constructor.

recordOutput :: RecordOutput

Creates a value of RecordOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

roOutputValue - The output value.

roOutputKey - The output key.

roDescription - The text description of the output.

roOutputValue :: Lens' RecordOutput (Maybe Text)

The output value.

roOutputKey :: Lens' RecordOutput (Maybe Text)

The output key.

roDescription :: Lens' RecordOutput (Maybe Text)

The text description of the output.

RecordTag

data RecordTag

A tag associated with the record, stored as a key-value pair.

See: recordTag smart constructor.

recordTag :: RecordTag

Creates a value of RecordTag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtValue - The value for this tag.

rtKey - The key for this tag.

rtValue :: Lens' RecordTag (Maybe Text)

The value for this tag.

rtKey :: Lens' RecordTag (Maybe Text)

The key for this tag.

ResourceDetail

data ResourceDetail

Detailed resource information.

See: resourceDetail smart constructor.

resourceDetail :: ResourceDetail

Creates a value of ResourceDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rARN - ARN of the resource.

rCreatedTime - Creation time of the resource.

rName - Name of the resource.

rId - Identifier of the resource.

rDescription - Description of the resource.

rARN :: Lens' ResourceDetail (Maybe Text)

ARN of the resource.

rCreatedTime :: Lens' ResourceDetail (Maybe UTCTime)

Creation time of the resource.

rName :: Lens' ResourceDetail (Maybe Text)

Name of the resource.

rId :: Lens' ResourceDetail (Maybe Text)

Identifier of the resource.

rDescription :: Lens' ResourceDetail (Maybe Text)

Description of the resource.

Tag

data Tag

Key-value pairs to associate with this provisioning. These tags are entirely discretionary and are propagated to the resources created in the provisioning.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - The ProvisioningArtifactParameter.TagKey parameter from DescribeProvisioningParameters .

tagValue - The desired value for this key.

tagKey :: Lens' Tag Text

The ProvisioningArtifactParameter.TagKey parameter from DescribeProvisioningParameters .

tagValue :: Lens' Tag Text

The desired value for this key.

TagOptionDetail

data TagOptionDetail

The TagOption details.

See: tagOptionDetail smart constructor.

tagOptionDetail :: TagOptionDetail

Creates a value of TagOptionDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

todValue - The TagOptionDetail value.

todActive - The TagOptionDetail active state.

todKey - The TagOptionDetail key.

todId - The TagOptionDetail identifier.

todValue :: Lens' TagOptionDetail (Maybe Text)

The TagOptionDetail value.

todActive :: Lens' TagOptionDetail (Maybe Bool)

The TagOptionDetail active state.

todKey :: Lens' TagOptionDetail (Maybe Text)

The TagOptionDetail key.

todId :: Lens' TagOptionDetail (Maybe Text)

The TagOptionDetail identifier.

TagOptionSummary

data TagOptionSummary

The TagOption summary key-value pair.

See: tagOptionSummary smart constructor.

tagOptionSummary :: TagOptionSummary

Creates a value of TagOptionSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tosValues - The TagOptionSummary value.

tosKey - The TagOptionSummary key.

tosValues :: Lens' TagOptionSummary [Text]

The TagOptionSummary value.

tosKey :: Lens' TagOptionSummary (Maybe Text)

The TagOptionSummary key.

UpdateProvisioningParameter

data UpdateProvisioningParameter

The parameter key-value pair used to update a ProvisionedProduct object. If UsePreviousValue is set to true, Value is ignored and the value for Key is kept as previously set (current value).

See: updateProvisioningParameter smart constructor.

updateProvisioningParameter :: UpdateProvisioningParameter

Creates a value of UpdateProvisioningParameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uppValue - The value to use for updating the product provisioning. Any constraints on this value can be found in the ProvisioningArtifactParameter parameter for Key .

uppKey - The ProvisioningArtifactParameter.ParameterKey parameter from DescribeProvisioningParameters .

uppUsePreviousValue - If true, uses the currently set value for Key , ignoring UpdateProvisioningParameter.Value .

uppValue :: Lens' UpdateProvisioningParameter (Maybe Text)

The value to use for updating the product provisioning. Any constraints on this value can be found in the ProvisioningArtifactParameter parameter for Key .

uppKey :: Lens' UpdateProvisioningParameter (Maybe Text)

The ProvisioningArtifactParameter.ParameterKey parameter from DescribeProvisioningParameters .

uppUsePreviousValue :: Lens' UpdateProvisioningParameter (Maybe Bool)

If true, uses the currently set value for Key , ignoring UpdateProvisioningParameter.Value .

UsageInstruction

data UsageInstruction

Additional information provided by the administrator.

See: usageInstruction smart constructor.

usageInstruction :: UsageInstruction

Creates a value of UsageInstruction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiValue - The usage instruction value for this type.

uiType - The usage instruction type for the value.

uiValue :: Lens' UsageInstruction (Maybe Text)

The usage instruction value for this type.

uiType :: Lens' UsageInstruction (Maybe Text)

The usage instruction type for the value.