Network.AWS.ServiceCatalog.UpdateConstraint

Description

Updates an existing constraint.

Creating a Request

updateConstraint

Arguments

:: Text
ucId

-> UpdateConstraint

Creates a value of UpdateConstraint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

ucDescription - The updated text description of the constraint.

ucId - The identifier of the constraint to update.

data UpdateConstraint

See: updateConstraint smart constructor.

Request Lenses

ucAcceptLanguage :: Lens' UpdateConstraint (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

ucDescription :: Lens' UpdateConstraint (Maybe Text)

The updated text description of the constraint.

ucId :: Lens' UpdateConstraint Text

The identifier of the constraint to update.

Destructuring the Response

updateConstraintResponse

Arguments

:: Int
ucrsResponseStatus

-> UpdateConstraintResponse

Creates a value of UpdateConstraintResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucrsStatus - The status of the current request.

ucrsConstraintDetail - The resulting detailed constraint information.

ucrsConstraintParameters - The resulting updated constraint parameters.

ucrsResponseStatus - -- | The response status code.

data UpdateConstraintResponse

See: updateConstraintResponse smart constructor.

Response Lenses

ucrsStatus :: Lens' UpdateConstraintResponse (Maybe RequestStatus)

The status of the current request.

ucrsConstraintDetail :: Lens' UpdateConstraintResponse (Maybe ConstraintDetail)

The resulting detailed constraint information.

ucrsConstraintParameters :: Lens' UpdateConstraintResponse (Maybe Text)

The resulting updated constraint parameters.

ucrsResponseStatus :: Lens' UpdateConstraintResponse Int

- | The response status code.