Network.AWS.ServiceCatalog.UpdateProduct

Description

Updates an existing product.

Creating a Request

updateProduct

Arguments

:: Text
upId

-> UpdateProduct

Creates a value of UpdateProduct with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upRemoveTags - Tags to remove from the existing list of tags associated with the product.

upOwner - The updated owner of the product.

upSupportURL - The updated support URL for the product.

upDistributor - The updated distributor of the product.

upName - The updated product name.

upAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

upAddTags - Tags to add to the existing list of tags associated with the product.

upSupportEmail - The updated support email for the product.

upDescription - The updated text description of the product.

upSupportDescription - The updated support description for the product.

upId - The identifier of the product for the update request.

data UpdateProduct

See: updateProduct smart constructor.

Request Lenses

upRemoveTags :: Lens' UpdateProduct [Text]

Tags to remove from the existing list of tags associated with the product.

upOwner :: Lens' UpdateProduct (Maybe Text)

The updated owner of the product.

upSupportURL :: Lens' UpdateProduct (Maybe Text)

The updated support URL for the product.

upDistributor :: Lens' UpdateProduct (Maybe Text)

The updated distributor of the product.

upName :: Lens' UpdateProduct (Maybe Text)

The updated product name.

upAcceptLanguage :: Lens' UpdateProduct (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

upAddTags :: Lens' UpdateProduct [Tag]

Tags to add to the existing list of tags associated with the product.

upSupportEmail :: Lens' UpdateProduct (Maybe Text)

The updated support email for the product.

upDescription :: Lens' UpdateProduct (Maybe Text)

The updated text description of the product.

upSupportDescription :: Lens' UpdateProduct (Maybe Text)

The updated support description for the product.

upId :: Lens' UpdateProduct Text

The identifier of the product for the update request.

Destructuring the Response

updateProductResponse

Arguments

:: Int
ursResponseStatus

-> UpdateProductResponse

Creates a value of UpdateProductResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ursProductViewDetail - The resulting detailed product view information.

ursTags - Tags associated with the product.

ursResponseStatus - -- | The response status code.

data UpdateProductResponse

See: updateProductResponse smart constructor.

Response Lenses

ursProductViewDetail :: Lens' UpdateProductResponse (Maybe ProductViewDetail)

The resulting detailed product view information.

ursTags :: Lens' UpdateProductResponse [Tag]

Tags associated with the product.

ursResponseStatus :: Lens' UpdateProductResponse Int

- | The response status code.