Network.AWS.ServiceCatalog.UpdateProvisionedProduct

Description

Requests updates to the configuration of an existing ProvisionedProduct object. If there are tags associated with the object, they cannot be updated or added with this operation. Depending on the specific updates requested, this operation may update with no interruption, with some interruption, or replace the ProvisionedProduct object entirely.

You can check the status of this request using the DescribeRecord operation.

Creating a Request

updateProvisionedProduct

Arguments

:: Text
uppUpdateToken

-> UpdateProvisionedProduct

Creates a value of UpdateProvisionedProduct with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uppProvisionedProductName - The updated name of the ProvisionedProduct object. Specify either ProvisionedProductName or ProvisionedProductId , but not both.

uppProvisioningArtifactId - The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

uppAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

uppPathId - The identifier of the path to use in the updated ProvisionedProduct object. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

uppProvisioningParameters - A list of ProvisioningParameter objects used to update the ProvisionedProduct object.

uppProvisionedProductId - The identifier of the ProvisionedProduct object to update. Specify either ProvisionedProductName or ProvisionedProductId , but not both.

uppProductId - The identifier of the ProvisionedProduct object.

uppUpdateToken - The idempotency token that uniquely identifies the provisioning update request.

data UpdateProvisionedProduct

See: updateProvisionedProduct smart constructor.

Request Lenses

uppProvisionedProductName :: Lens' UpdateProvisionedProduct (Maybe Text)

The updated name of the ProvisionedProduct object. Specify either ProvisionedProductName or ProvisionedProductId , but not both.

uppProvisioningArtifactId :: Lens' UpdateProvisionedProduct (Maybe Text)

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

uppAcceptLanguage :: Lens' UpdateProvisionedProduct (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

uppPathId :: Lens' UpdateProvisionedProduct (Maybe Text)

The identifier of the path to use in the updated ProvisionedProduct object. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

uppProvisioningParameters :: Lens' UpdateProvisionedProduct [UpdateProvisioningParameter]

A list of ProvisioningParameter objects used to update the ProvisionedProduct object.

uppProvisionedProductId :: Lens' UpdateProvisionedProduct (Maybe Text)

The identifier of the ProvisionedProduct object to update. Specify either ProvisionedProductName or ProvisionedProductId , but not both.

uppProductId :: Lens' UpdateProvisionedProduct (Maybe Text)

The identifier of the ProvisionedProduct object.

uppUpdateToken :: Lens' UpdateProvisionedProduct Text

The idempotency token that uniquely identifies the provisioning update request.

Destructuring the Response

updateProvisionedProductResponse

Arguments

:: Int
upprsResponseStatus

-> UpdateProvisionedProductResponse

Creates a value of UpdateProvisionedProductResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upprsRecordDetail - The detailed result of the UpdateProvisionedProduct request, containing the inputs made to that request, the current state of the request, a pointer to the ProvisionedProduct object that the request is modifying, and a list of any errors that the request encountered.

upprsResponseStatus - -- | The response status code.

data UpdateProvisionedProductResponse

See: updateProvisionedProductResponse smart constructor.

Response Lenses

upprsRecordDetail :: Lens' UpdateProvisionedProductResponse (Maybe RecordDetail)

The detailed result of the UpdateProvisionedProduct request, containing the inputs made to that request, the current state of the request, a pointer to the ProvisionedProduct object that the request is modifying, and a list of any errors that the request encountered.

upprsResponseStatus :: Lens' UpdateProvisionedProductResponse Int

- | The response status code.