Network.AWS.ServiceCatalog.UpdateProvisioningArtifact

Description

Updates an existing provisioning artifact's information. This operation does not work on a provisioning artifact associated with a product that has been shared with you.

Creating a Request

updateProvisioningArtifact

Arguments

:: Text
upaProductId

-> Text
upaProvisioningArtifactId

-> UpdateProvisioningArtifact

Creates a value of UpdateProvisioningArtifact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upaName - The updated name of the provisioning artifact.

upaAcceptLanguage - The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

upaDescription - The updated text description of the provisioning artifact.

upaProductId - The product identifier.

upaProvisioningArtifactId - The identifier of the provisioning artifact for the update request. This is sometimes referred to as the product version.

data UpdateProvisioningArtifact

See: updateProvisioningArtifact smart constructor.

Request Lenses

upaName :: Lens' UpdateProvisioningArtifact (Maybe Text)

The updated name of the provisioning artifact.

upaAcceptLanguage :: Lens' UpdateProvisioningArtifact (Maybe Text)

The language code to use for this operation. Supported language codes are as follows: "en" (English) "jp" (Japanese) "zh" (Chinese) If no code is specified, "en" is used as the default.

upaDescription :: Lens' UpdateProvisioningArtifact (Maybe Text)

The updated text description of the provisioning artifact.

upaProductId :: Lens' UpdateProvisioningArtifact Text

The product identifier.

upaProvisioningArtifactId :: Lens' UpdateProvisioningArtifact Text

The identifier of the provisioning artifact for the update request. This is sometimes referred to as the product version.

Destructuring the Response

updateProvisioningArtifactResponse

Arguments

:: Int
uparsResponseStatus

-> UpdateProvisioningArtifactResponse

Creates a value of UpdateProvisioningArtifactResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uparsStatus - The status of the current request.

uparsInfo - Additional information about the provisioning artifact update request.

uparsProvisioningArtifactDetail - The resulting detailed provisioning artifact information.

uparsResponseStatus - -- | The response status code.

data UpdateProvisioningArtifactResponse

See: updateProvisioningArtifactResponse smart constructor.

Response Lenses

uparsStatus :: Lens' UpdateProvisioningArtifactResponse (Maybe RequestStatus)

The status of the current request.

uparsInfo :: Lens' UpdateProvisioningArtifactResponse (HashMap Text Text)

Additional information about the provisioning artifact update request.

uparsProvisioningArtifactDetail :: Lens' UpdateProvisioningArtifactResponse (Maybe ProvisioningArtifactDetail)

The resulting detailed provisioning artifact information.

uparsResponseStatus :: Lens' UpdateProvisioningArtifactResponse Int

- | The response status code.