Network.AWS.ServiceCatalog.UpdateTagOption

Description

Updates an existing TagOption.

Creating a Request

updateTagOption

Arguments

:: Text
utoId

-> UpdateTagOption

Creates a value of UpdateTagOption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utoValue - The updated value.

utoActive - The updated active state.

utoId - The identifier of the constraint to update.

data UpdateTagOption

See: updateTagOption smart constructor.

Request Lenses

utoValue :: Lens' UpdateTagOption (Maybe Text)

The updated value.

utoActive :: Lens' UpdateTagOption (Maybe Bool)

The updated active state.

utoId :: Lens' UpdateTagOption Text

The identifier of the constraint to update.

Destructuring the Response

updateTagOptionResponse

Arguments

:: Int
utorsResponseStatus

-> UpdateTagOptionResponse

Creates a value of UpdateTagOptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utorsTagOptionDetail - The resulting detailed TagOption information.

utorsResponseStatus - -- | The response status code.

data UpdateTagOptionResponse

See: updateTagOptionResponse smart constructor.

Response Lenses

utorsTagOptionDetail :: Lens' UpdateTagOptionResponse (Maybe TagOptionDetail)

The resulting detailed TagOption information.

utorsResponseStatus :: Lens' UpdateTagOptionResponse Int

- | The response status code.